fix: 修复 CI 四类失败（lint/格式/配置校验/看板集成测试）
- CI 工作流: golangci-lint-action v6→v7（v6 不支持 golangci-lint v2）；
  deploy-lint 与 smoke 的 .env 补齐 DASHBOARD_PASSWORD/SESSION_SECRET，
  compose 三种形态插值校验恢复通过
- Go: gofmt 全量格式化；修复 lint 暴露的 43 个存量问题
  （errorlint 错误包装 / govet 变量遮蔽 / contextcheck×3 显式豁免 /
  nilerr 对账取消返回 ctx.Err() / ineffassign 死赋值 / ST1005 等）
- 看板: ruff format；新增 tests/__init__.py 消除 mypy 重复模块名；
  修复 4 处类型错误；SQL 列名跟进多 provider 改名
  （volc_key_id→upstream_key_id, volc_keys→upstream_keys）——
  此前看板全部查询因列名失配而查不到数据，11 个集成测试必挂

+++ internal/gateway/gateway_test.go
@@ -1267,7 +1267,7 @@ func TestCount_上游不回usage时流水仍记实扣量(t *testing.T) {
 	if len(recs) != 1 {
 		t.Fatalf("流水条数 = %d, 期望 1", len(recs))
 	}
-	if got := int64(recs[0].CountUnits); got != used {
+	if got := recs[0].CountUnits; got != used {
 		t.Errorf("流水 count_units = %d, 实扣 = %d —— 账面与配额不一致", got, used)
 	}
 }
